+++ docs/changelog.md
@@ -1,7 +1,13 @@
 ## Changelog
 Reverse chronological order.
+### Week 2 (Sept 8-14, 2024)
+#### Added
+- NIST framework research (NIST-Research.md)
+- Added scanner for passwords and account lockout. (scanner.py)
 
-### Week 1 (Sept 5-15, 2024)
+
+
+### Week 1 (Sept 1-7, 2024)
 #### Added
 - Initial project structure (Architecture.md)
 - NIST framework research (No commit)